DDA-86: Move getUntrackedDiograph to utils for future re-use
