+++ src/react/utils/index.js
@@ -1,3 +1,4 @@
 export { debounce } from './debounce'
 export { useGetImageUrl } from './useGetImageUrl'
 export { convertRelativePath } from './convertRelativePath'
+export { getUntrackedDiograph } from './getUntrackedDiograph'
